Proper algorithm implementation.
This commit is based on the previous collaboration commit, but cleans
the implementation up through a major refactoring of the code.

Furthermore, this commit introduces proper support for all precedence
rules in the algorithm as well as an implementation for taking care of
the second query parameter (class string).

Lastly, this commit fixes memory leaks introduced previously.