Hide struct definition and introduce converter functions.
For a cleaner API and to avoid breakage upon implementation changes, we
move the struct definition back from the public API to the internal
definitions. This patch also introduces two functions to retrieve the
(string) value of a resource as well as convert a resource value into an
integer.

fixes #3