Update parser and algorithms for spec-conform model.
This commit changes the model xcb-util-xrm uses to represent entries.
Previously, we considered both '?' and '*' (separate) components. However,
only '?' is actually a component on its own, while '*' is just a type of
binding between components (much like '.').

For example, this also means that while 'a*b' is correct, 'a?b' is not and
must rather be 'a.?.b' (or 'a*?.b' etc.).

Therefore, we clean up our internal model and correctly distinguish between
the binding type and a "normal" component versus a wildcard. This requires
both a change in the parser and in the algorithm implementation.

On top, this commit introduced proper validation of the characters used
in component names, conforming to the spec. Lastly, we clean up the tests
here and make them easier to read.