Ignore lines starting with '!' and '#'.
We skip lines that are comments or directives. We have to expect these in
case the caller wants to parse their own string (otherwise, xrdb took care
of removing comments already).

While ignoring comments is safe, a caller may expect include directives
to be handled. However, at least for now this is not supported.