Handle ':' in the beginning of a resource explicitly.
This resolves a leftover TODO comment. Back when it was written, it
was assumed that a ':' would be a valid character in a resource name.
However, this is not the case, so we can simply reject in this
situation. A test case for this already existed, but this makes the
reject more explicit.