Reworked the entire API.
This commit changes the API from the ground up (of course, in a breaking
way). We don't actually have much of a reason to use a "context" since
most of what we do does not directly concern X (or xcb for that matter).
The only case where we need a X connection is when fetching the resource
manager content from the root window.

Therefore, we move from having a context object as the central concept of
this library to using the database struct for this. This is closer related
to how the Xlib equivalent works and will also allow a more straight-forward
API when introducing the APIs to combine / merge / manipulate databases.