Don't export symbols that are actually internal functions.
This commit renames (almost, see below) all functions to have a
"__" prefix unless they are intended to be in the API. This is
necessary to avoid exporting those symbols accidentally. Thanks
to @psychon (Uli Schlachter) for noticing this.

There currently are two exceptions to this, which are
 - xcb_xrm_entry_parse
 - xcb_xrm_entry_free
We currently have to export those symbols as they are used to
unit-test the parser.

fixes #50