Fix "make distcheck"
Automake supports out of tree builds (mkdir build && cd build && ../configure).
"make distcheck" tests if this really works. In this case, the tests are run
with their working directory set to the build directory (the place where
binaries are built (more or less)). However, the file test in
tests/tests_database.c wants to read files in the source directory. The path to
this directory is passed on by automake in $srcdir.

This commit makes the test use this variable to figure out paths. If $srcdir is
not set, then "." is assumed. That way, tests can still be called by hand.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>