Automatically run tests_database under Xvfb
tests_database is the only test requiring an X11 server. Also, that server must
not have a RESOURCE_MANAGER set on its root window, else the test will fail.

The easiest way to make this work portably is to have a small wrapper script
that uses xvfb-run to run the test. If xvfb-run is not found, the test is
skipped.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>