Build tests conditionally
This moves the check for x11 and x11-xcb to configure.ac. The tests are only
enabled if the necessary libraries are found. Additionally, this commit adds
--enable-tests / --disable-tests flags to configure to forcefully disable tests
and, more important, to make configure fail if the dependencies for the tests
could not be found.

I hope someone can tell me a nicer way to implement this...

Fixes: https://github.com/Airblader/xcb-util-xrm/issues/60
Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>