Parse entries with empty value correctly.
This patch fixes how an entry like

    some.component:

is parsed. Previously, this would error out and ignore the entry,
but the correct behavior is to parse it and assign the empty value
to it. This is needed for applications to differentiate between a
resource not being defined and it having an empty value.

fixes #71